requirements.txt
# Requirements for Service Quality Oracle
# Requires Python 3.9+

# Core dependencies
pyyaml==6.0.2

# Configuration management
tomli==2.2.1  # TOML support for Python < 3.11

# Scheduling and resilience
schedule==1.2.2
pytz==2025.2
tenacity==8.5.0

# Google Cloud BigQuery for data processing
google-cloud-bigquery==3.26.0
bigframes==1.42.0
ibis-framework==10.5.0

# Data processing and validation
pandas==2.2.3
pandera==0.20.4
numpy>=2.0.0  # Added as pandas dependency

# Database interaction
psycopg2-binary==2.9.10
SQLAlchemy==1.4.54

# Blockchain integration
web3==6.16.0
eth-account>=0.11.0  # Explicit dependency for web3

# GraphQL and subgraph integration (for future subgraph functionality)
gql==3.5.2
requests-toolbelt==1.0.0

# HTTP and API
requests==2.32.3
aiohttp>=3.9.0  # For async HTTP requests (used by web3)

# Testing
pytest==8.3.3

# Development tools
ruff==0.6.8
pip==25.1
